fix timming function
git-svn-id: http://psp.jim.sh/svn/ps2ware/trunk@333 5c23000f-aefb-0310-8550-8024f1b2be0e

+++ quake1_ps2/sys.h
@@ -54,7 +54,7 @@ void Sys_Printf (char *fmt, ...);
 
 void Sys_Quit (void);
 
-double Sys_FloatTime (void);
+float Sys_FloatTime (void);
 
 char *Sys_ConsoleInput (void);
